Start converting to python3
The latest version of django-ldapdb includes changes to byte/string
handling that make switching to a Python3 runtime easier than fighting
with the library. This is probably a good thing in general anyway. We
could try to use six and maintain a 2/3 compat codebase, but just
cutting over to python3 *should* be easier.