Set ALLOWED_HOSTS to ['*'] by default
Set Django's ALLOWED_HOSTS setting to ['*'] by default. This effectively
disables the internal Django Host header validation mechanism even when
DEBUG is false. Usage in this form should only be considered safe in
a production environment if an HTTP proxy is between the end user and
the app and that proxy is only routing specific Host values to the app.
This should always be the case in Wikimedia Labs and production
environments where nginx and/or varnish will be used in front of the
application.

For extra protection, production deployments are encouraged to set
a more restrictive ALLOWED_HOSTS value in the locally deployed
/etc/striker/striker.ini configuration file.

Change-Id: Ifc1952ffc1ea23d5c156b468c76d64e2d76ac290