Allow configuration of static_root
Add ini setting to change the path of the collected static files
directory. This can be used in a production environment to place the
collected static files in a location that makes more sense for serving
with a traditional web server like apache or nginx. The Django runtime
still needs to have access to this directory as it contains the manifest
file providing filenames for the long cache duration files.

Change-Id: I416a501d2950cbdc295b586468cecec1540f6f59