Add pretty error pages
Add error page handler views and templates to replace the default Django
error reporting.

A new library, django-log-request-id` is also added to tag all log
messages with a per-request identifier. This identifier is can be used
to lookup and correlate log events that occur in a single web
request/response cycle. The id is not passed on to downstream
processors, nor is it persisted in any way that can be used to correlate
multiple requests (i.e. user tracking).

For ease of testing, the following routes are also created:
* /e400
* /e403
* /e404 (not actually created, but symmetrical for testing 404 output)
* /e500

Bug: T143949
Change-Id: I3169fbcbb46953602f499bdd82a749c800edd0a6