Add account creation initial screen
Add a new "register" app that will handle account creation. This is just
a stub at this point, but it should give a reasonable idea of where I'm
headed with the feature. A wizard style multi-step process will collect
information from the user, validate it, and eventually add a new LDAP
user account.

The TOU warning is taken directly from Special:CreateAccount on wikitech.

Note: the addition of new css styles here means that Django's
collectstatic process needs to be run when this is tested/deployed.

Bug: T144710
Change-Id: Ic8f5171a1ed43585def3d9ffe5b92f5295ca4f1c