Collect data needed to create a new LDAP account
Collect the username, shell account name, email address and password
needed to create a new LDAP account. This does not yet create the
account. It also doesn't validate that the username and shell account
name are available.

This change introduces a new Python library (django-formtools). Virtual
environments will need to be updated for testing and wheels will need to
be updated after merging before it can be deployed to production.

Bug: T144710
Change-Id: I47e76a228d5eeacf540c0b1d49e0f4472ca26196