Create LDAP and Striker users from registration form data
Use the data collected by the registration form wizard to create a new
LDAP record. When that succeeds, add a Striker user model that ties the
new LDAP record to the SUL account.

The LDAP record created is follows the same core logic as the
combination of LdapAuthentication and OpenStackManager used on Wikitech:
- ObjectClass attributes: person, inetOrgPerson, organizationalPerson,
  ldapPublicKey, posixAccount, shadowAccount
- Cn and sn attributes both set to the wiki/gerrit/apache username.
- Uid attribute set to unix shell name.
- Numeric uid chosen by adding one to the current max posixAccount uid
  in LDAP.
- Numeric uid allocation retried once in case of collision due to race
  with another account creation.
- Numeric gid set to 500 (configurable).
- Login shell set to /bin/bash (configurable).
- Home directory set to /home/{uid}
- Password attribute added with no client side hashing.

The existing PosixGroup LDAP model has also been updated to use field
naming that follows the LDAP attribute naming. This convention should
make working with LDAP models easier than remembering some arbitrary
remapping of the attribute names.

Bug: T144710
Change-Id: Ib09262ab6fdb8d6d490e436c4a27cb60d2beafb7