Validate new usernames with action=query&list=users&usprop=cancreate
Call the Action API on Wikitech to validate that a desired username is
not blacklisted via Titleblacklist or other means. This is much easier
than recreating all of the blacklisting logic in Striker itself and
trying to somehow keep it in sync with the Wikitech and Wikimedia global
rules.

This patch introduces a new dependency on the mwclient python library
which is used to talk to the Action API. Full production deployment will
require an associated update to the labs/striker/wheels.git repository.

Bug: T147024
Change-Id: Id79568880efe84bbd12c9cc130cdc5fc9c20541d