Add support for authenticated Action API use
Upgrade to mwoauth >= 0.8.2 and add configuration for authenticating to
the MediaWiki Action API using OAuth. A method for getting an anonymous
client is also provided and used when checking for username availability
so that on wiki rights do not accidentally allow usernames that would
otherwise be blocked.

When this is deployed to production there will be private Puppet changes
needed to add configuration settings for the StrikerBot wikitech user.

Bug: T144712
Change-Id: Ibfc4b35498e59bf02da3c8562fea802a573f139c