Check and enforce OATH account protection
Query MediaWiki when a user has successfully authenticated using their
LDAP password to see if the Wikitech account has two-factor auth
protection enabled. If enabled flag the authenticated user as needing
OATH token validation to continue. A new Django middleware has been
added that will check for the OATH flag and related validation data
during the initial handling of each request. When OATH is needed but has
not been validated the request will be redirected to a form that
collects the token and validates it via the MediaWiki Action API.

Validation results are stored in the session and include the time of the
validation which may be used in the future to provide enhanced security
checks for sensitive operations but forcing a new token validation if
too much time has passed.

Use of the OATH check and validation API endpoints requires a MediaWiki
user account that has the `oathauth-api-all` right.

Bug: T144712
Change-Id: I46d5b419254ded6aec6aea6148ef8f642b11e76e