Fix account creation bug
Apparently I did a really poor job of testing the account creation flow
with some of the changes that came as a result of code review. The
LABSAUTH_MAX_UID setting was missing and UID and GID settings values were not
properly cast to integers which caused problems in computing the next
available value.

Change-Id: Ic83cd166cb365f7e0e58095da782f3ccf82942cd