Add toolinfo.json style data
Add a ToolInfo model to collect and store a superset of data needed to
create toolinfo.json descriptions of tools. Additional fields are added
for license, issue tracker, and documentation. The current models do not
support recording keywords.

Basic display and editing of ToolInfo models is integrated with existing
tool detail views. Users who are not members of a ToolInfo's associated
tool account are allowed to edit a subset of the model's fields.

ToolInfo models are versioned using django-reversion. A view is provided
for viewing history of a given ToolInfo. Tool members can revert any
edit to a prior version and also suppress display of revisions. I'm not
trying to recreate the wiki I promise!

Bug: T149458
Change-Id: Ibe0bbada144c164e41f5bcf3e9b186fd2757dabb