Make django's makemigrations happy
The ldapdb models are not actually managed by Django's migrations
system, but makemigrations apparently doesn't know that. Rather than
adding a new migration file that does nothing, I'm manually backporting
the changes it wants to make to the initial migration for the project.
This is a complete no-op for both development and production.

Change-Id: I46335e5464d72da175c7ef2402ee968062cbf949