Tools LDAP model migrations
Django really wants to manage all the models it can find. This doesn't
work out very well with the LDAP backend so we need to remember to mark
LDAP models as unmanaged. I'm not really sure why they don't do this in
the base class. Even when marked as unmanaged Django wants to have
a no-op migration that tell it that the models exist.

Change-Id: Ie38d4fabb275665837ad23e5cdf0e05f07a8a14c