Change ToolInfo authors to list of strings
Using tool maintainers/LDAP users as the author list was a bad idea, so
lets undo it. Tools can have origins from lots of places and we should
not assume that everything will be credited to active local users.

The django migrations for this may make a mess of things. I ended up
needing to manually change the foreign key constraints in the
tools_toolinfo_authors table. This should not be a problem in
a production environment where the tables are managed with hand curated
SQL statements. If it turns out to be problematic for developers we may
need to hack up the migration files in a followup.

Bug: T149458
Change-Id: I586148a9dcfd3ef7d6b1f09e64b890a695dc4e8e