Check tools.admin membership for admin rights
Use a direct check of LDAP for admin rights rather than the is_staff bit
in the local user object. These two checks are technically functionally
equivalent, but the latter is only updated at login which means that it
could be false positive or negative depending on changes that have been
made in LDAP since the session started.

Change-Id: I1333bb694c6a33667978fe8528ae7a71ffe93d4c