striker.tools.models.Maintainer: fix __str__
The 'username' property was renamed to 'cn' in aaf2a47, but the __str__
method was not updated at the same time. Strangely this seems to have
been the cause of a bug where only some eligible maintainers could be
added to tools. I'm still a bit confused about how validation was
working for some accounts and not for others while this typo existed.

Bug: T173963
Change-Id: I2b5e20e910b69e8015ff74b04c9b25bb02e76055