Temporarily replace requirements.txt with `pip --freeze` from prod
Django has released a new major version. We have pinned Django to
a range, but libraries we are using without version pinning are being
updated to be compatible with the breaking changes from that version.
This in turn means that a clean install using `pip install -r
requirements.txt` (as is done in our CI environment) is picking up
incompatible versions of some libraries.

I'm not happy about such strict pinning as it makes testing newer
releases a bit harder, but for the moment this is the nicest way to get
our CI environment and other new installs back to a sane state. This
pinning should be loosened and the package license comments restored as
part of the Django 1.11 LTS upgrade (T192487).

Bug: T198076
Change-Id: I8a4d9fb23fe80acf4ce28bca2bdd95b90b7eb511
Depends-On: Iff97e7671ba9a1ab6041372cc3cb69c2242dfc7b