Upgrade to Django 2.1
Upgrade Django and other tracked libraries to latest stable versions.
There are a lot of breaking changes to deal with as a result of this,
but managing that seems easier than finding all of the version pinning
needed in order to remain in the Django 1.x series. At this point the
latest LTS version (1.11) only gives us a few more months of support
than the 2.1 series.

Bug: T192487
Bug: T199488
Bug: T212398
Change-Id: I559bda812ac5a20d3ce90d88772858f07a2d006d