Switch diffusion.repository.search search field
The 'query' constraint is documented as "Find objects matching
a fulltext search query. See "Search User Guide" in the documentation
for details." In my local testing it appears to take "a while" (1-5
seconds) before whatever index backs the 'query' lookup to actually
be populated and able to match the repo. The 'shortNames' constraint
seems to be indexed immediately.

Bug: T182142
Change-Id: If0138b2f383bb85aed02a2f1b2cd9d1247aee81f