Add CEE log formatter and handler
Add a slim wrapper around logstash.LogstashFormatterVersion1 that
prepends the "@cee:" cookie which is used by rsyslog's message
normalization module to distinguish json and non-json log records.

This new formatter and a corresponding handler are added to the logging
configuration. Use of this handler can be enabled for a deployment by
setting the logging.HANDLERS ini value to "cee".

Bug: T217932
Change-Id: If3519da8f8b9ca127443de00f79bc7827bcee981