Diffusion: change options for new repos
Update the options we set for new repos to:
* Disable publishing of commits
* Remove legacy "autoclose" setting

Autoclose was hard deprecated upstream in Phabricator with disable of
publishing providing the historic behavior of preventing commit messages
from closing referenced tasks. This has a bit of an unfortunate side
effect of also preventing commits from being mentioned in the timeline
of referenced tasks. This loss of functionality is pretty minor however,
and can be changed by the repo owner(s) manually if desired.

Bug: T227508
Change-Id: Ibec48e54108e784ec2fe4f41ed76fed5ea3cddea