Bump to Django 2.2.3 + other library upgrades
Update our minimum Django version to 2.2.3. The 2.2 series is an LTS
version, which means we will in theory be able to stick with it for
quite some time. In practice, we will have to see. It will depend
a bit on the level of breaking change in the 2.3 series. As we have seen
in the past it can be difficult to untangle all of the cross library
dependencies when a lot of breaking change is happening in the Django
core.

Change-Id: I91ca1225eb2b996d40c2358b7e74802da2c8369b