Guard against empty/null names when doing Phabricator lookups
I stumbled across yet another way to blow up the Phabricator lookups.
A user might not have any Phabricator or MediaWiki associations locally,
and this can lead to a null being inserted in the lookup list. The
built-in filter() function does a nice job of stripping those out before
calling the Phabricator backend.

Change-Id: I36af4a4e62119fb7528d3a6c60065347891cd306