Fork django-ldapdb to workaround T228332
The root cause of T228332 has been traced to our upgrade to django-ldapdb
1.3.0. A bug [0] and associated pull request [1] have been sent upstream to
fix the problem with our use of the library. While we wait for that PR to
land and a new version of django-ldapdb to be released we will switch to
using the fork of django-ldapdb where the PR was implemented.

[0]: https://github.com/django-ldapdb/django-ldapdb/issues/196
[1]: https://github.com/django-ldapdb/django-ldapdb/pull/197

Bug: T228332
Change-Id: I1f4c890fa4049d70b90a7a54b2670a95619c9496