dev: Move things out of docker-compose.yaml
Move as many upstream container customizations as possible out to our
docker-compose.yaml file. These customizations are moved into Dockerfile
files in contrib/docker. This will make it easier to reuse the
customized containers without docker-compose.yaml when desired. The main
use case for this is expected to be a staging environment in a Cloud VPS
project.

Bug: T306096
Change-Id: I735fc89032ed998081f72e83faefa1cc92377460