Add 'whitenoise' for static file serving
Add the WhiteNoise (<http://whitenoise.evans.io/>) library and related
configuration for serving static assets. WhiteNoise provides a static
file service implemented in Python that is performant enough to be used
in a production environment. This works best in coordination with a CDN
that can provide long term caching of versioned static assets. In the
Wikimedia production environment the Wikimedia edge CDN should fill this
purpose.

Bug: T306469
Change-Id: Id4e4d5884497edeaa839457b9fbdd5db06a33292