fix: Lookup GitLab repos by id rather than name
We will continue to use names as labels in the UI and also to link to
GitLab via the original name of the repo at its time of creation, but we
will only lookup data on the repo by numeric id to protect against
issues arising from renaming on the GitLab side. We will also update the
local name for a repo if it is found to have changed while rendering
a repo detail screen.

Bug: T317377
Change-Id: Iff054d9a73e5d4e7e8d22e666a7b78e9cb4d31a3