feat: Add UI for disable/enable of tool accounts
Add "disable tool" button to a tool's detail page that can be used by
tool maintainers and Toolforge admins to mark a tool as disabled in
LDAP. Once a tool has been marked as disabled either via Striker or by
external means, show the elapsed time since disable. For maintainers and
admins, also show a button to re-enable the tool.

Bug: T285403
Change-Id: I757aa2111cef8e9ad32b7e19c184830e0f58cddf