dev: limit history from `make tail`
Add a `--tail=1000` argument to our call to docker-compose's logs action
so that only the most recent 1000 lines of each container's logs are
shown. The gitlab container makes a huge amount of logging. So much so
that tailing the full log when running `make tail` can take many minutes
for little value.

Change-Id: I5f229e9b0887df6cd23f1da5f6174e05ac301b89