fix: replace revision.views._RollBackRevisionView
Replace the revision.views._RollBackRevisionView subclass of Exception
that was removed upstream with our own local Exception subclass. This
use of exceptions for flow control is a bit suspicious in it's own
right, but I'm not sure that reworking the whole view right now is
a good use of time.

Bug: T325622
Change-Id: I9e7288b5afc08a7a9e32dad02c9cc00497671921