dev(keystone): Hide blocked LDAP users from Keystone
Add a `user_filter` configuration directive to our Keystone service that
will hide LDAP users that have been blocked on ldapwiki.local from
Keystone. This is accomplished by requiring that users do not have
a pwdPolicySubentry attribute pointing to the cn=disabled policy that is
applied to lock LDAP accounts.

Change-Id: I020ba010fd9cc5c2c05e240a3fc2dae907816c64