dev(docker): Temporarily use quay.io/bd808/bitnami/phabricator:2021
We do not yet have a permanent replacement for the now deleted
docker.io/bitnami/phabricator:2021 image. For now we can use the copy of
it that was saved to quay.io/bd808/bitnami/phabricator:2021 by Bryan
Davis. This feels slightly better than leaving the broken upstream
reference in the config and hoping that folks will find the Phabricator
task describing the workaround.

Bug: T340080
Change-Id: Ic8a5b07e93740ce68d2e7b13d2bd87257045da98